+++ src/clients.rs
@@ -1,3 +1,4 @@
-pub use self::cli::Cli;
+pub use self::{cli::Cli, http::Http};
 
 mod cli;
+mod http;