+++ src/core.rs
@@ -1,12 +1,18 @@
 pub use self::{
     container::Container,
+    container_async::ContainerAsync,
     docker::{Docker, Logs, Ports, RunArgs},
     image::{Image, Port, WaitFor},
     wait_for_message::{WaitError, WaitForMessage},
 };
 
 mod container;
+mod container_async;
 mod docker;
 pub mod env;
 mod image;
 mod wait_for_message;
+
+pub(crate) mod logs;
+
+pub(crate) use container_async::DockerAsync;